hwp_mcp/hwpx_vision/requirements.txt
mcp>=1.2.0
pydantic>=2.5
httpx>=0.27
markdown-it-py>=3.0
pymupdf>=1.24
python-hwpx>=2.9
pillow>=10.0
lxml>=5.1
